Changed id type in models to Long for consistency

+++ app/src/main/java/com/example/uberapp_tim/model/message/Message.java
@@ -5,14 +5,14 @@ import com.example.uberapp_tim.model.users.User;
 import java.time.LocalDateTime;
 
 public class Message {
-    private int id;
+    private Long id;
     private String msgText;
     private LocalDateTime timeOfSending;
     private MessageType messageType;
     private User sender;
     private User receiver;
 
-    public Message(int id, String msgText, LocalDateTime timeOfSending, MessageType messageType,
+    public Message(Long id, String msgText, LocalDateTime timeOfSending, MessageType messageType,
                    User receiver, User sender) {
         this.id = id;
         this.msgText = msgText;
